- refactor : 음식점 세부사항 호출 서비스 로직 이름 변경
