+++ StudentSoup/src/main/java/ProjectDoge/StudentSoup/controller/restaurant/RestaurantDetailCallController.java
@@ -2,7 +2,7 @@ package ProjectDoge.StudentSoup.controller.restaurant;
 
 
 import ProjectDoge.StudentSoup.dto.restaurant.RestaurantDetailRequestDto;
-import ProjectDoge.StudentSoup.service.restaurantmenu.RestaurantMenuCallService;
+import ProjectDoge.StudentSoup.service.restaurant.RestaurantDetailCallService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Pageable;
@@ -17,12 +17,12 @@ import java.util.concurrent.ConcurrentHashMap;
 @RestController
 public class RestaurantDetailCallController {
 
-    private final RestaurantMenuCallService restaurantMenuCallService;
+    private final RestaurantDetailCallService restaurantDetailCallService;
 
     @PostMapping("/{restaurantId}")
     public ConcurrentHashMap<String, Object> callRestaurantDetail(@PathVariable Long restaurantId,
                                                                   @RequestBody RestaurantDetailRequestDto dto,
                                                                   @PageableDefault(size = 4) Pageable pageable){
-        return restaurantMenuCallService.restaurantDetailCall(dto.getRestaurantId(), dto.getMemberId(), pageable);
+        return restaurantDetailCallService.restaurantDetailCall(dto.getRestaurantId(), dto.getMemberId(), pageable);
     }
 }
